feat(endpoints): show materialization runs modal in configuration tab
Reuses the SQL editor's materialization status & runs modal on the
EndpointScene Configuration tab so users can trigger syncs and inspect
run history without leaving the endpoint. The materialization UI is
extracted from QueryInfo into MaterializationStatusPanel +
MaterializationStatusModal, so both surfaces render identical controls
backed by a single materializationJobsLogic.

Generated-By: PostHog Code
Task-Id: e0712d75-03a7-41e2-9c52-951e3fc5eef4
