+++ frontend/src/scenes/data-warehouse/editor/output-pane-tabs/QueryInfo.tsx
@@ -1,33 +1,19 @@
 import { useActions, useValues } from 'kea'
 
-import { IconRefresh, IconRevert, IconTarget, IconX } from '@posthog/icons'
-import { LemonDialog, LemonTable, Link, Spinner } from '@posthog/lemon-ui'
+import { IconTarget } from '@posthog/icons'
+import { LemonTable, Link, Spinner } from '@posthog/lemon-ui'
 
 import { FEATURE_FLAGS } from 'lib/constants'
-import { dayjsUtcToTimezone } from 'lib/dayjs'
 import { LemonButton } from 'lib/lemon-ui/LemonButton'
-import { LemonProgress } from 'lib/lemon-ui/LemonProgress'
 import { LemonSegmentedButton } from 'lib/lemon-ui/LemonSegmentedButton'
-import { LemonSelect } from 'lib/lemon-ui/LemonSelect'
-import { LemonTag, LemonTagType } from 'lib/lemon-ui/LemonTag'
+import { LemonTag } from 'lib/lemon-ui/LemonTag'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { humanFriendlyDetailedTime, humanFriendlyDuration, humanFriendlyNumber } from 'lib/utils'
+import { humanFriendlyDetailedTime } from 'lib/utils'
 import { dataWarehouseViewsLogic } from 'scenes/data-warehouse/saved_queries/dataWarehouseViewsLogic'
-import { LogsViewer } from 'scenes/hog-functions/logs/LogsViewer'
-import { teamLogic } from 'scenes/teamLogic'
-import { userLogic } from 'scenes/userLogic'
+import { MaterializationStatusPanel } from 'scenes/data-warehouse/saved_queries/MaterializationStatusPanel'
 
-import {
-    DataModelingJob,
-    DataWarehouseSavedQuery,
-    DataModelingSyncInterval,
-    LineageNode,
-    LogEntryLevel,
-    OrNever,
-} from '~/types'
-
-const LOG_LEVELS: LogEntryLevel[] = ['LOG', 'INFO', 'WARN', 'WARNING', 'ERROR']
+import { DataWarehouseSavedQuery, LineageNode } from '~/types'
 
 import { UpstreamGraph } from '../sidebar/graph/UpstreamGraph'
 import { sqlEditorLogic } from '../sqlEditorLogic'
@@ -38,127 +24,18 @@ interface QueryInfoProps {
     view?: DataWarehouseSavedQuery | null
 }
 
-function getMaterializationStatusMessage(
-    rowsMaterialized: number,
-    progressPercentage: number,
-    rowsExpected: number
-): string {
-    const percentComplete = Math.round(Math.min(100, (rowsMaterialized / rowsExpected) * 100))
-    switch (true) {
-        case rowsMaterialized === 0:
-            return `Spinning up spikes — starting materialization job... ${percentComplete}% complete.`
-        case progressPercentage < 10:
-            return `Digging into SQL... executing your query now... ${percentComplete}% complete.`
-        case progressPercentage < 25:
-            return `First ${humanFriendlyNumber(rowsMaterialized)} rows tucked away... ${percentComplete}% complete.`
-        case progressPercentage < 50:
-            return `${humanFriendlyNumber(rowsMaterialized)} rows shipped to storage... ${percentComplete}% complete.`
-        case progressPercentage < 90:
-            return `Still going — ${humanFriendlyNumber(
-                rowsMaterialized
-            )} rows written... ${percentComplete}% complete.`
-        case progressPercentage === 100:
-            return `Wrapping up — ${humanFriendlyNumber(
-                rowsMaterialized
-            )} rows processed... ${percentComplete}% complete.`
-        default:
-            return `Almost there — ${humanFriendlyNumber(
-                rowsMaterialized
-            )} rows processed... ${percentComplete}% complete.`
-    }
-}
-
-const OPTIONS = [
-    {
-        value: 'never' as OrNever,
-        label: ' No resync',
-    },
-    {
-        value: '15min' as DataModelingSyncInterval,
-        label: ' Resync every 15 mins',
-    },
-    {
-        value: '30min' as DataModelingSyncInterval,
-        label: ' Resync every 30 mins',
-    },
-    {
-        value: '1hour' as DataModelingSyncInterval,
-        label: ' Resync every 1 hour',
-    },
-    {
-        value: '6hour' as DataModelingSyncInterval,
-        label: ' Resync every 6 hours',
-    },
-    {
-        value: '12hour' as DataModelingSyncInterval,
-        label: ' Resync every 12 hours',
-    },
-    {
-        value: '24hour' as DataModelingSyncInterval,
-        label: ' Resync Daily',
-    },
-    {
-        value: '7day' as DataModelingSyncInterval,
-        label: ' Resync Weekly',
-    },
-    {
-        value: '30day' as DataModelingSyncInterval,
-        label: ' Resync Monthly',
-    },
-]
-
-function getMaterializationDisabledReasons(
-    currentJobStatus: string | null,
-    startingMaterialization: boolean
-): {
-    sync: string | false
-    cancel: string | false
-    revert: string | false
-} {
-    return {
-        sync:
-            currentJobStatus === 'Running'
-                ? 'Materialization is already running'
-                : startingMaterialization
-                  ? 'Materialization is starting'
-                  : false,
-        cancel: currentJobStatus !== 'Running' ? 'Materialization is not running' : false,
-        revert: currentJobStatus === 'Running' ? 'Cannot revert while materialization is running' : false,
-    }
-}
-
 export function QueryInfo({ tabId, view }: QueryInfoProps): JSX.Element {
     const { editingView, upstream, upstreamViewMode } = useValues(sqlEditorLogic)
     const targetView = view ?? editingView
     const infoLogic = infoTabLogic({ tabId, viewId: targetView?.id })
     const { sourceTableItems } = useValues(infoLogic)
-    const { runDataWarehouseSavedQuery, saveAsView, setUpstreamViewMode } = useActions(sqlEditorLogic)
+    const { saveAsView, setUpstreamViewMode } = useActions(sqlEditorLogic)
     const { featureFlags } = useValues(featureFlagLogic)
-    const { timezone } = useValues(teamLogic)
-    const { user } = useValues(userLogic)
-    const showDebugLogs = user?.is_staff || user?.is_impersonated
 
     const isLineageDependencyViewEnabled = featureFlags[FEATURE_FLAGS.LINEAGE_DEPENDENCY_VIEW]
-    const isDagSchedulesOnly = !!featureFlags[FEATURE_FLAGS.DATA_MODELING_BACKEND_V2]
 
-    const { dataModelingJobs, dataModelingJobsLoading, hasMoreJobsToLoad, startingMaterialization } =
-        useValues(infoLogic)
-    const { loadDataModelingJobs, loadOlderDataModelingJobs, setStartingMaterialization } = useActions(infoLogic)
-
-    const { dataWarehouseSavedQueryMapById, updatingDataWarehouseSavedQuery, initialDataWarehouseSavedQueryLoading } =
+    const { updatingDataWarehouseSavedQuery, initialDataWarehouseSavedQueryLoading } =
         useValues(dataWarehouseViewsLogic)
-    const {
-        updateDataWarehouseSavedQuery,
-        cancelDataWarehouseSavedQuery,
-        materializeDataWarehouseSavedQuery,
-        revertMaterialization,
-    } = useActions(dataWarehouseViewsLogic)
-
-    // note: targetView is stale, but dataWarehouseSavedQueryMapById gets updated
-    const savedQuery = targetView ? dataWarehouseSavedQueryMapById[targetView.id] : null
-
-    const currentJobStatus = dataModelingJobs?.results?.[0]?.status || null
-    const { sync, cancel, revert } = getMaterializationDisabledReasons(currentJobStatus, startingMaterialization)
 
     if (initialDataWarehouseSavedQueryLoading) {
         return (
@@ -171,287 +48,36 @@ export function QueryInfo({ tabId, view }: QueryInfoProps): JSX.Element {
     return (
         <div className="overflow-auto" data-attr="sql-editor-sidebar-query-info-pane">
             <div className="flex flex-col flex-1 gap-4">
-                <div>
-                    <div className="flex flex-row items-center gap-2">
-                        <h3 className="mb-0">Materialization</h3>
-                        <LemonTag type="warning">BETA</LemonTag>
-                        {savedQuery?.latest_error && savedQuery.status === 'Failed' && (
-                            <Tooltip title={savedQuery.latest_error} interactive>
-                                <LemonTag type="danger">Error</LemonTag>
-                            </Tooltip>
-                        )}
-                    </div>
+                {targetView ? (
+                    <MaterializationStatusPanel viewId={targetView.id} />
+                ) : (
                     <div>
-                        {savedQuery?.is_materialized ? (
-                            <div>
-                                {savedQuery?.last_run_at ? (
-                                    `Last run at ${humanFriendlyDetailedTime(savedQuery?.last_run_at)}`
-                                ) : (
-                                    <div>
-                                        <span>Materialization scheduled</span>
-                                    </div>
-                                )}
-                                <div className="flex gap-4 mt-2">
-                                    <LemonButton
-                                        className="whitespace-nowrap"
-                                        loading={startingMaterialization || currentJobStatus === 'Running'}
-                                        disabledReason={sync}
-                                        onClick={() => {
-                                            if (targetView) {
-                                                setStartingMaterialization(true)
-                                                runDataWarehouseSavedQuery(targetView.id)
-                                            }
-                                        }}
-                                        type="secondary"
-                                        sideAction={{
-                                            icon: <IconX fontSize={16} />,
-                                            tooltip: 'Cancel materialization',
-                                            onClick: () => targetView && cancelDataWarehouseSavedQuery(targetView.id),
-                                            disabledReason: cancel,
-                                        }}
-                                    >
-                                        {startingMaterialization
-                                            ? 'Starting...'
-                                            : currentJobStatus === 'Running'
-                                              ? 'Running...'
-                                              : 'Sync now'}
-                                    </LemonButton>
-                                    {!isDagSchedulesOnly && (
-                                        <LemonSelect
-                                            className="h-9"
-                                            disabledReason={sync}
-                                            value={
-                                                targetView
-                                                    ? dataWarehouseSavedQueryMapById[targetView.id]?.sync_frequency ||
-                                                      'never'
-                                                    : 'never'
-                                            }
-                                            onChange={(newValue) => {
-                                                if (targetView && newValue) {
-                                                    updateDataWarehouseSavedQuery({
-                                                        id: targetView.id,
-                                                        sync_frequency: newValue,
-                                                        types: [[]],
-                                                        lifecycle: 'update',
-                                                    })
-                                                }
-                                            }}
-                                            loading={updatingDataWarehouseSavedQuery}
-                                            options={OPTIONS}
-                                        />
-                                    )}
-                                    {targetView && (
-                                        <LemonButton
-                                            type="secondary"
-                                            size="small"
-                                            tooltip="Revert materialized view to view"
-                                            disabledReason={revert}
-                                            icon={<IconRevert />}
-                                            onClick={() => {
-                                                LemonDialog.open({
-                                                    title: 'Revert materialization',
-                                                    maxWidth: '30rem',
-                                                    description:
-                                                        'Are you sure you want to revert this materialized view to a regular view? This will stop all future materializations and remove the materialized table. You will always be able to go back to a materialized view at any time.',
-                                                    primaryButton: {
-                                                        status: 'danger',
-                                                        children: 'Revert materialization',
-                                                        onClick: () => revertMaterialization(targetView.id),
-                                                    },
-                                                    secondaryButton: {
-                                                        children: 'Cancel',
-                                                    },
-                                                })
-                                            }}
-                                        />
-                                    )}
-                                </div>
-                            </div>
-                        ) : (
-                            <div>
-                                <p className="text-xs">
-                                    Materialized views are a way to pre-compute data in your data warehouse. This allows
-                                    you to run queries faster and more efficiently.
-                                    <br />
-                                    <Link
-                                        data-attr="materializing-help"
-                                        to="https://posthog.com/docs/data-warehouse/views#materializing-and-scheduling-a-view"
-                                        target="_blank"
-                                    >
-                                        Learn more about materialization
-                                    </Link>
-                                    .
-                                </p>
-                                <LemonButton
-                                    size="small"
-                                    onClick={() => {
-                                        if (targetView) {
-                                            materializeDataWarehouseSavedQuery(targetView.id)
-                                        } else {
-                                            saveAsView({ materializeAfterSave: true })
-                                        }
-                                    }}
-                                    type="primary"
-                                    loading={updatingDataWarehouseSavedQuery}
-                                >
-                                    {targetView ? 'Materialize' : 'Save and materialize'}
-                                </LemonButton>
-                            </div>
-                        )}
-                    </div>
-                </div>
-                {savedQuery && (
-                    <>
-                        <div className="flex items-start justify-between">
-                            <div>
-                                <h3>Materialization Runs</h3>
-                                <p className="text-xs">
-                                    The last runs for this materialized view. These can be scheduled or run on demand.
-                                </p>
-                            </div>
-                            <LemonButton
-                                icon={<IconRefresh />}
-                                size="small"
-                                type="secondary"
-                                onClick={() => loadDataModelingJobs(savedQuery.id)}
-                                loading={dataModelingJobsLoading}
-                                disabledReason={startingMaterialization ? 'Materialization is starting' : undefined}
-                                tooltip="Refresh runs"
-                            />
+                        <div className="flex flex-row items-center gap-2">
+                            <h3 className="mb-0">Materialization</h3>
+                            <LemonTag type="warning">BETA</LemonTag>
                         </div>
-                        <LemonTable
+                        <p className="text-xs">
+                            Materialized views are a way to pre-compute data in your data warehouse. This allows you to
+                            run queries faster and more efficiently.
+                            <br />
+                            <Link
+                                data-attr="materializing-help"
+                                to="https://posthog.com/docs/data-warehouse/views#materializing-and-scheduling-a-view"
+                                target="_blank"
+                            >
+                                Learn more about materialization
+                            </Link>
+                            .
+                        </p>
+                        <LemonButton
                             size="small"
-                            loading={dataModelingJobsLoading && !dataModelingJobs?.results?.length}
-                            dataSource={dataModelingJobs?.results || []}
-                            columns={[
-                                {
-                                    title: 'Status',
-                                    dataIndex: 'status',
-                                    render: (_, job: DataModelingJob) => {
-                                        const { status, error, rows_materialized, rows_expected } = job
-                                        const statusToType: Record<string, LemonTagType> = {
-                                            Completed: 'success',
-                                            Failed: 'danger',
-                                            Running: 'warning',
-                                        }
-                                        const type = statusToType[status] || 'warning'
-
-                                        const progressPercentage =
-                                            rows_expected && rows_expected > 0
-                                                ? Math.min(100, (rows_materialized / rows_expected) * 100)
-                                                : 0
-
-                                        // Only show progress if there is > 0 progress and we have expected rows
-                                        // many small result sets will never show progress as they are written in only 1 batch
-                                        if (status === 'Running' && progressPercentage > 0 && rows_expected !== null) {
-                                            return (
-                                                <Tooltip
-                                                    placement="right"
-                                                    title={getMaterializationStatusMessage(
-                                                        rows_materialized,
-                                                        progressPercentage,
-                                                        rows_expected
-                                                    )}
-                                                >
-                                                    <div className="w-[68px]">
-                                                        <LemonProgress percent={progressPercentage} />
-                                                    </div>
-                                                </Tooltip>
-                                            )
-                                        }
-
-                                        return error && status !== 'Completed' ? (
-                                            <Tooltip title={error} interactive>
-                                                <LemonTag type={type}>{status}</LemonTag>
-                                            </Tooltip>
-                                        ) : (
-                                            <LemonTag type={type}>{status}</LemonTag>
-                                        )
-                                    },
-                                },
-                                {
-                                    title: 'Rows',
-                                    dataIndex: 'rows_materialized',
-                                    render: (_, { rows_materialized, status }: DataModelingJob) =>
-                                        (status === 'Running' || status === 'Cancelled') && rows_materialized === 0
-                                            ? '~'
-                                            : humanFriendlyNumber(rows_materialized),
-                                },
-                                {
-                                    title: 'Updated',
-                                    dataIndex: 'last_run_at',
-                                    render: (_, { last_run_at }: DataModelingJob) =>
-                                        humanFriendlyDetailedTime(last_run_at),
-                                },
-                                {
-                                    title: 'Duration',
-                                    render: (_, job: DataModelingJob) => {
-                                        if (job.status === 'Running') {
-                                            return 'In progress'
-                                        }
-                                        // Convert date strings to timestamps before subtraction
-                                        const start = new Date(job.created_at).getTime()
-                                        const end = new Date(job.last_run_at).getTime()
-
-                                        if (start > end) {
-                                            return 'N/A'
-                                        }
-
-                                        return humanFriendlyDuration((end - start) / 1000)
-                                    },
-                                },
-                            ]}
-                            expandable={
-                                dataModelingJobs?.results?.length && savedQuery
-                                    ? {
-                                          expandedRowRender: (job: DataModelingJob) => (
-                                              <div className="p-4">
-                                                  <LogsViewer
-                                                      logicKey={`data_modeling_run:${job.id}`}
-                                                      sourceType="data_modeling_run"
-                                                      sourceId={savedQuery.id}
-                                                      groupByInstanceId={false}
-                                                      hideDateFilter
-                                                      hideLevelsFilter
-                                                      hideInstanceIdColumn
-                                                      defaultFilters={{
-                                                          instanceId: job.workflow_run_id,
-                                                          dateFrom: dayjsUtcToTimezone(job.created_at, timezone).format(
-                                                              'YYYY-MM-DD HH:mm:ss'
-                                                          ),
-                                                          dateTo: job.last_run_at
-                                                              ? dayjsUtcToTimezone(job.last_run_at, timezone)
-                                                                    .add(1, 'hour')
-                                                                    .format('YYYY-MM-DD HH:mm:ss')
-                                                              : undefined,
-                                                          levels: showDebugLogs ? ['DEBUG', ...LOG_LEVELS] : LOG_LEVELS,
-                                                      }}
-                                                  />
-                                              </div>
-                                          ),
-                                          rowExpandable: () => true,
-                                          noIndent: true,
-                                      }
-                                    : undefined
-                            }
-                            nouns={['run', 'runs']}
-                            emptyState="No runs available"
-                            footer={
-                                hasMoreJobsToLoad && (
-                                    <div className="flex items-center m-2">
-                                        <LemonButton
-                                            center
-                                            fullWidth
-                                            onClick={() => loadOlderDataModelingJobs()}
-                                            loading={dataModelingJobsLoading}
-                                        >
-                                            Load older runs
-                                        </LemonButton>
-                                    </div>
-                                )
-                            }
-                        />
-                    </>
+                            onClick={() => saveAsView({ materializeAfterSave: true })}
+                            type="primary"
+                            loading={updatingDataWarehouseSavedQuery}
+                        >
+                            Save and materialize
+                        </LemonButton>
+                    </div>
                 )}
                 {!isLineageDependencyViewEnabled && (
                     <>
