+++ frontend/src/scenes/data-warehouse/editor/output-pane-tabs/infoTabLogic.ts
@@ -1,18 +1,11 @@
-import { actions, afterMount, connect, kea, key, listeners, path, props, propsChanged, reducers, selectors } from 'kea'
-import { loaders } from 'kea-loaders'
-import { subscriptions } from 'kea-subscriptions'
+import { connect, kea, key, listeners, path, props, selectors } from 'kea'
 
-import api, { PaginatedResponse } from 'lib/api'
 import { databaseTableListLogic } from 'scenes/data-management/database/databaseTableListLogic'
 import { dataWarehouseViewsLogic } from 'scenes/data-warehouse/saved_queries/dataWarehouseViewsLogic'
-
-import { DataModelingJob } from '~/types'
+import { materializationJobsLogic } from 'scenes/data-warehouse/saved_queries/materializationJobsLogic'
 
 import { sqlEditorLogic } from '../sqlEditorLogic'
 import type { infoTabLogicType } from './infoTabLogicType'
-
-const REFRESH_INTERVAL = 10000
-const DEFAULT_JOBS_PAGE_SIZE = 10
 
 export interface InfoTableRow {
     name: string
@@ -34,66 +27,15 @@ export const infoTabLogic = kea<infoTabLogicType>([
     connect((props: InfoTabLogicProps) => ({
         values: [
             sqlEditorLogic({ tabId: props.tabId }),
-            ['metadata', 'editingView'],
+            ['metadata'],
             databaseTableListLogic,
             ['posthogTablesMap', 'dataWarehouseTablesMap'],
             dataWarehouseViewsLogic,
             ['dataWarehouseSavedQueryMap'],
         ],
         actions: [sqlEditorLogic({ tabId: props.tabId }), ['loadUpstream']],
+        logic: props.viewId ? [materializationJobsLogic({ viewId: props.viewId })] : [],
     })),
-    actions({
-        setStartingMaterialization: (starting: boolean) => ({ starting }),
-    }),
-    loaders(({ values }) => ({
-        dataModelingJobs: [
-            null as PaginatedResponse<DataModelingJob> | null,
-            {
-                loadDataModelingJobs: async (savedQueryId: string) => {
-                    return await api.dataWarehouseSavedQueries.dataWarehouseDataModelingJobs.list(
-                        savedQueryId,
-                        values.dataModelingJobs?.results.length
-                            ? Math.max(values.dataModelingJobs.results.length, DEFAULT_JOBS_PAGE_SIZE)
-                            : DEFAULT_JOBS_PAGE_SIZE,
-                        0
-                    )
-                },
-                loadOlderDataModelingJobs: async () => {
-                    const nextUrl = values.dataModelingJobs?.next
-
-                    if (!nextUrl) {
-                        return values.dataModelingJobs
-                    }
-
-                    const res = await api.get<PaginatedResponse<DataModelingJob>>(nextUrl)
-                    res.results = [...(values.dataModelingJobs?.results ?? []), ...res.results]
-
-                    return res
-                },
-            },
-        ],
-    })),
-    reducers({
-        startingMaterialization: [
-            false,
-            {
-                setStartingMaterialization: (_, { starting }: { starting: boolean }) => starting,
-                loadDataModelingJobsSuccess: (
-                    state: boolean,
-                    { dataModelingJobs }: { dataModelingJobs: PaginatedResponse<DataModelingJob> | null }
-                ) => {
-                    const currentJobStatus = dataModelingJobs?.results?.[0]?.status
-                    if (
-                        currentJobStatus &&
-                        ['Running', 'Completed', 'Failed', 'Cancelled'].includes(currentJobStatus)
-                    ) {
-                        return false
-                    }
-                    return state
-                },
-            },
-        ],
-    }),
     selectors({
         sourceTableItems: [
             (s) => [s.metadata, s.dataWarehouseSavedQueryMap],
@@ -124,33 +66,18 @@ export const infoTabLogic = kea<infoTabLogicType>([
                 )
             },
         ],
-        hasMoreJobsToLoad: [(s) => [s.dataModelingJobs], (dataModelingJobs) => !!dataModelingJobs?.next],
     }),
-    afterMount(({ actions, props }) => {
-        if (props.viewId) {
-            actions.loadDataModelingJobs(props.viewId)
+    listeners(({ actions, props }) => {
+        if (!props.viewId) {
+            return {}
+        }
+        const jobsLogic = materializationJobsLogic({ viewId: props.viewId })
+        return {
+            [jobsLogic.actionTypes.loadDataModelingJobsSuccess]: () => {
+                if (props.viewId) {
+                    actions.loadUpstream(props.viewId)
+                }
+            },
         }
     }),
-    propsChanged(({ actions, props }, oldProps) => {
-        if (props.viewId && props.viewId !== oldProps.viewId) {
-            actions.loadDataModelingJobs(props.viewId)
-        }
-    }),
-    listeners(({ actions, cache }) => ({
-        loadDataModelingJobsSuccess: ({ payload }) => {
-            if (payload) {
-                actions.loadUpstream(payload)
-            }
-
-            cache.disposables.add(() => {
-                const timeoutId = setTimeout(() => {
-                    if (payload) {
-                        actions.loadDataModelingJobs(payload)
-                    }
-                }, REFRESH_INTERVAL)
-                return () => clearTimeout(timeoutId)
-            }, 'dataModelingJobsRefreshTimeout')
-        },
-    })),
-    subscriptions(() => ({})),
 ])
